refactor: update SearchBar import paths to common directory
