+++ app/(tabs)/_layout.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Tabs } from 'expo-router';
 import { Ionicons } from '@expo/vector-icons';
 
-import Header from '@/components/Header';
+import Header from '@/components/common/Header';
 import { useColorScheme } from 'react-native';
